==> Notebooks/1-data-cleaning.ipynb <==
import pandas as pd
import requests
import re

# CSV files were generated using webscraper.io to pull metadata from every speaker in the Speech Accent Archive on February 17th, 2023
df1 = pd.read_csv('speech-accent-archive-part-1.csv')
df2 = pd.read_csv('speech-accent-archive-part-2.csv')
df = pd.concat([df1, df2], axis=0).reset_index()

# Dropping rows with native language not specified
df = df.dropna(subset=['native-language'])

# Creating cleaned data columns using regex
df['url-id'] = df.apply(lambda x: re.findall('=(\d+)', x['web-scraper-start-url'])[0], axis=1)
df['length-residence2'] = df.apply(lambda x: re.findall('length of english residence: (\d+\.*\d*) years', x['length-residence'])[0], axis=1)
df['english-residence2'] = df.apply(lambda x: re.findall('english residence: (.*)', x['english-residence'])[0] if len(re.findall('english residence: (.*)', x['english-residence'])) > 0 else "", axis=1)
df['age-of-english-onset2'] = df.apply(lambda x: re.findall('age of english onset: (\d*)', x['age-of-english-onset'])[0], axis=1)
df['other-languages2'] = df.apply(lambda x: re.findall('other language\(s\): (.*)', x['other-languages'])[0] if len(re.findall('other language\(s\): (.*)', x['other-languages'])) > 0 else "", axis=1)
df['birth-place2'] = df.apply(lambda x: re.findall('birth place: (.*) \(map\)', x['birth-place'])[0] if len(re.findall('birth place: (.*) \(map\)', x['birth-place'])) > 0 else "", axis=1)
df['age2'] = df.apply(lambda x: re.findall('age, sex: (\d+),', x['age-sex'])[0] if len(re.findall('age, sex: (\d+),', x['age-sex'])) > 0 else pd.NA, axis=1)
df['sex2'] = df.apply(lambda x: re.findall('age, sex: \d+, (.*)', x['age-sex'])[0] if len(re.findall('age, sex: \d+, (.*)', x['age-sex'])) > 0 else pd.NA, axis=1)
df['english-dialect'] = df['birth-place2']
df['english-dialect'] = df.apply(lambda x: x['birth-place2'] if x['native-language'] == 'english' else "", axis=1)
df['english-dialect'] = df.apply(lambda x: 'usa' if 'usa' in x['english-dialect'] else x['english-dialect'], axis=1)
df['english-dialect'] = df.apply(lambda x: 'uk' if 'uk' in x['english-dialect'] else x['english-dialect'], axis=1)
df['english-dialect'] = df.apply(lambda x: 'canada' if 'canada' in x['english-dialect'] else x['english-dialect'], axis=1)
df['english-dialect'] = df.apply(lambda x: 'ireland' if 'ireland' in x['english-dialect'] else x['english-dialect'], axis=1)
df['english-dialect'] = df.apply(lambda x: 'new zealand' if 'new zealand' in x['english-dialect'] else x['english-dialect'], axis=1)
df['english-dialect'] = df.apply(lambda x: 'jamaica' if 'jamaica' in x['english-dialect'] else x['english-dialect'], axis=1)
df['english-dialect'] = df.apply(lambda x: 'india' if 'india' in x['english-dialect'] else x['english-dialect'], axis=1)
df['english-dialect'] = df.apply(lambda x: 'south africa' if 'south africa' in x['english-dialect'] else x['english-dialect'], axis=1)

df['native-language'].value_counts()[:30]

# Creating the first version of the dataset, containing the 10 most frequent native languages, English broken down into dialects

march = df[df['native-language'].isin(['english', 'spanish', 'arabic', 'mandarin', 'korean', 'french', 'russian', 'portuguese', 'dutch'])]
def accent_class(x):
    if x['native-language'] == 'english':
        if x['english-dialect'] == 'usa':
            return 'usa'
        if x['english-dialect'] == 'uk':
            return 'uk'
        else:
            return pd.NA
    else:
        return x['native-language']

march['accent'] = march.apply(accent_class, axis=1)

march['accent'].value_counts()

march = march.dropna(subset=['accent'])
march['age'] = march['age2']
march['sex'] = march['sex2']
march = march[['url-id', 'age', 'sex', 'accent']]
march = march.reset_index()
march.to_csv('speech-accent-archive-march-version.csv')
march

# Downloading the corresponding mp3 files from the web

saa_url = "https://accent.gmu.edu/"

with requests.Session() as req:

    for index, row, in df.iterrows():
        mp3_url = row['mp3-link']
        name = row['url-id']

        download = req.get(saa_url + mp3_url)
        if download.status_code == 200:
            with open(name + ".mp3", 'wb') as f:
                f.write(download.content)
        else:
            print(f"Download Failed For File {name}")
