==> Notebooks/4-proficiency-scores.ipynb <==
import pandas as pd
import numpy as np

## Calculating language proficiency scores

df = pd.read_csv("speech-accent-archive-processed2.csv")

df[['url-id',
       'length-residence2', 'english-residence2', 'age-of-english-onset2',
       'other-languages2', 'birth-place2', 'age2', 'sex2', 'english-dialect']]

# The goal of this step is to identify which speakers belong to the test and validation sets as I did not save that information in this csv
# I am reconstructing it from the folder structure where the data is stored

import os

def collect_filenames(path):
    filenames = []

    for root, dirs, files in os.walk(path):
        for file in files:
            file_without_ext, _ = os.path.splitext(file)
            filenames.append(file_without_ext)

    return filenames

path1 = "D:\\Data Science and Entrepreneurship MSc\\Thesis\\Data\\SAA\\s-f-o\\test"
path2 = "D:\\Data Science and Entrepreneurship MSc\\Thesis\\Data\\SAA\\s-f-o\\validation"
collected_filenames = collect_filenames(path1) + collect_filenames(path2)
print(collected_filenames)
print(len(collected_filenames))


def remove_trailing_zero(filenames):
    return [filename[:-2] if filename.endswith('_0') else filename for filename in filenames]

filenames_without_trailing_zero = remove_trailing_zero(collected_filenames)
print(filenames_without_trailing_zero)

# Convert the filenames to integers
filenames_as_integers = [int(filename) for filename in filenames_without_trailing_zero]

# Filter the DataFrame such that only test and validation speakers are shown
filtered_df = df[df['url-id'].isin(filenames_as_integers)]

filtered_df


labels_df = pd.read_csv("speech-accent-archive-march-version.csv")
labels_df.rename(columns={'accent': 'label'}, inplace=True)
labels_df

# Merge the DataFrames on the 'url-id' column
merged_df = filtered_df.merge(labels_df[['url-id', 'label']], on='url-id', how='left')
#merged_df = df.merge(labels_df[['url-id', 'label']], on='url-id', how='left')

# Display the merged DataFrame
merged_df

# Making sure
df = merged_df[merged_df['label'].isin(['arabic', 'dutch', 'french', 'korean', 'mandarin', 'portuguese', 'russian', 'spanish'])]

# Calculating proficiency scores
# Normalized scores were calculated prior to filtering to only validation/test sets
# But normalized scores were not utilized in the final thesis version

A = 10  # log years of exposure
B = 15  # log years in country
#C = 100  # inverse age started learning
D = 1  # Weight for years learning English before 18
E = 0.2  # Weight for percentage of life in English-speaking country

# Calculate the years of exposure
df['years_of_exposure'] = df['age2'] - df['age-of-english-onset2']

# Calculate years learning English before 18
df['years_learning_before_18'] = df.apply(lambda x: min(18 - x['age-of-english-onset2'], x['years_of_exposure']) if x['age-of-english-onset2'] < 18 else 0, axis=1)

# Calculate percentage of life in English-speaking country
df['percentage_life_in_english_country'] = (df['length-residence2'] / df['age2']) * 100

# Calculate the proficiency score
#df['proficiency'] = (df['years_of_exposure'] * A) + (df['length-residence2'] * B) + (C / df['age-of-english-onset2']) + (df['years_learning_before_18'] * D) + (df['percentage_life_in_english_country'] * E)
df['proficiency'] = (np.log(df['years_of_exposure'] + 1) * A) + (np.log(df['length-residence2'] + 1) * B) + (df['years_learning_before_18'] * D) + (df['percentage_life_in_english_country'] * E)


# Filter out the person who started learning at age 0
df = df[df['proficiency']<1000]

# Normalize the proficiency scores to a range of 0-100
min_proficiency = df['proficiency'].min()
max_proficiency = df['proficiency'].max()
print(min_proficiency, max_proficiency)
df['normalized_proficiency'] = ((df['proficiency'] - min_proficiency) / (max_proficiency - min_proficiency)) * 100
df


prof_df = df[['url-id', 'length-residence2', 'english-residence2', 'age-of-english-onset2', 'birth-place2', 'age2', 'proficiency', 'normalized_proficiency', 'label']]
prof_df

# Group the data by 'label' and calculate the average 'normalized_proficiency' per group
average_normalized_proficiency_per_label = df.groupby('label')['proficiency'].mean()

# Convert the result to a DataFrame for better readability
average_normalized_proficiency_per_label_df = average_normalized_proficiency_per_label.reset_index()

# Rename the columns for clarity
average_normalized_proficiency_per_label_df.columns = ['label', 'average_normalized_proficiency']

# Display the result
average_normalized_proficiency_per_label_df


# Group the data by 'label' and calculate the average 'normalized_proficiency' per group
average_normalized_proficiency_per_label = df.groupby('label')['proficiency'].mean()

# Calculate the standard deviation of 'normalized_proficiency' per group
std_normalized_proficiency_per_label = df.groupby('label')['proficiency'].std()

# Convert the results to a DataFrame for better readability
average_normalized_proficiency_per_label_df = average_normalized_proficiency_per_label.reset_index()
std_normalized_proficiency_per_label_df = std_normalized_proficiency_per_label.reset_index()

# Merge the average and standard deviation DataFrames
result_df = pd.merge(average_normalized_proficiency_per_label_df, std_normalized_proficiency_per_label_df, on='label')

# Rename the columns for clarity
result_df.columns = ['label', 'average_normalized_proficiency', 'std_normalized_proficiency']

# Display the result
result_df


prof_df.to_csv('proficiency_test_df.csv')

prof_df['proficiency'].mean()

prof_df['proficiency'].std()

## Prepare proficiency score plot

# create this table:
# rows: each speaker from validation and test
# columns: id, label, proficiency
df = df[['url-id', 'label', 'proficiency']]

df

# create 5 proficiency bins
df['prof-group'] = pd.qcut(df['proficiency'], 5, labels=[1, 2, 3, 4, 5])

df['prof-group'].value_counts()

# save this table and feed it to model evaluator ----> the rest of the steps were performed in the model_evaluator notebook
df.to_csv('proficiency_groups.csv')

# create a new row: binary classified correcly (int)

# calculate mean and sd per bin and plot it
